+++ Chord Provider/ChordProParser/ChordProParser+processTextblock.swift
@@ -1,0 +1,34 @@
+//
+//  ChordProParser+processTextblock.swift
+//  Chord Provider
+//
+//  © 2025 Nick Berendsen
+//
+
+import Foundation
+
+extension ChordProParser {
+
+    // MARK: Process a texblock environment
+
+    /// Process a texblock environment
+    /// - Parameters:
+    ///   - text: The text to process
+    ///   - currentSection: The current ``Song/Section``
+    ///   - song: The whole ``Song``
+    static func processTextblock(
+        text: String,
+        currentSection: inout Song.Section,
+        song: inout Song
+    ) {
+        /// Start with a fresh line
+        let line = Song.Section.Line(
+            sourceLineNumber: song.lines,
+            environment: .textblock,
+            directive: .environmentLine,
+            source: text.trimmingCharacters(in: .whitespacesAndNewlines),
+            plain: text.trimmingCharacters(in: .whitespacesAndNewlines)
+        )
+        currentSection.lines.append(line)
+    }
+}