+++ Chord Provider/Extensions/NSFont+extension.swift
@@ -1,0 +1,10 @@
+//
+//  NSFont+extension.swift
+//  Chord Provider
+//
+//  © 2025 Nick Berendsen
+//
+
+import AppKit
+
+extension NSFont: @unchecked @retroactive Sendable {}