+++ Chord Provider/Extensions/NSParagraphStyle+extension.swift
@@ -1,0 +1,10 @@
+//
+//  NSParagraphStyle+extension.swift
+//  Chord Provider
+//
+//  © 2025 Nick Berendsen
+//
+
+import AppKit
+
+extension NSParagraphStyle: @unchecked @retroactive Sendable {}